requirements.txt
Django==1.8.11
django-celery==3.1.17
redis==2.10.5
MySQL-python==1.2.5
elasticsearch==2.3.0
gunicorn
py-zabbix
djangorestframework
requests
pykafka
pymongo
